interpreter: initial implementation for parse tree interpreter
The interpreter evaluates the passed parse tree recursively and procudes
a sum evaluation. `eval` is a high-order function that takes, as it's
first argument, a random number generator engine. Calling the function
will output the actual evaluator that takes the parse tree as an
argument.

The random number generator engine is used in the roll calculations. A
base pseudo number generator engine is provided in the `lib.rs`.