ast: initial implementation of the abstract syntax tree
Implements enum representation for expressions and operators, which
should produce the complete parse tree with a parser.

Add helper functions to easily create different expression types, like
binary roll expressions or numeric literals. The only purpose for these
helper functions is to make it more ergonomic and readable to define
a parse tree.