wasm: initial implementation wasm droll library
Wasm binary is initially meant to be used in browser and nodejs
environments, but possibly the build/publish steps will form to contain
additional environments too.

The wasm library exposes just one function `roll`, which takes the dice
notation as an input and returns the summed result of the rolls. It uses
the default droll interpreter.