droll: add `calc_roll` to simplify dice rolls in base interpreter
The function `calc_roll` simplifies the interpreter calls when calling
pseudo-random engine with recursive evulation. `calc_roll` is a high
order function that takes, first, the pseudo-random generator, and then
the number of dice and the number of sides.

This commit also fixes a rounding error when calling the base
pseudo-random engine. The rounding error resulted in never getting the
highest possible value of the roll. For example, a d6 would never
produce a 6.