Use pratt parsing instead of only recursion
Pratt parsing makes binding power and operator precedence simple. This
allows for faster iterations of new functionality for the parser.

The implementation of pratt parser was heavily inspired by this article:
https://matklad.github.io/2020/04/13/simple-but-powerful-pratt-parsing.html

Pratt parsing contains a loop that calls the parse function recursively
to get the right-hand side of the expression.

Each operator token gets a left and right binding power and if the left
binding power is lower than the given minimum binding power, we break
from the loop. We also break from the loop if we run out of tokens.

Before the loop we calculate the left-hand side expression by matching
the tokens. If we encounter an operator token we parse an unary
expression.